requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
packaging>=23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
